return fast if fixture is already installed/uninstalled and nothing to do

+++ spring-fixtures/src/main/java/net/savantly/spring/fixture/AbstractBaseFixture.java
@@ -27,6 +27,10 @@ public abstract class AbstractBaseFixture<T, R extends CrudRepository<T, ?>> imp
 	}
 
 	public void install() {
+		if(installed){
+			log.warn("Fixture already installed");
+			return;
+		}
 		log.info("Beginning Fixture Install");
 		synchronized (lock) {
 			ensureDependenciesAreInstalled();
@@ -51,6 +55,10 @@ public abstract class AbstractBaseFixture<T, R extends CrudRepository<T, ?>> imp
 	}
 
 	public void uninstall() {
+		if(!installed){
+			log.warn("Fixture not installed");
+			return;
+		}
 		log.info("Beginning Fixture Uninstall");
 		synchronized (lock) {
 			repository.delete(entityList);
